refactor: OPTIC-429: Step 2 to avoid complexity in reading the PR
