+++ src/components/MainView/GridView/GridView.styl
