+++ src/components/MainView/index.js
@@ -1,3 +1,1 @@
-import { DataView as DataViewOld } from "./DataViewOld/Table";
-
-export const DataView = DataViewOld;
+export { DataView } from "./DataView/Table";
